extract route logic into sub route
