+++ deployables/app/app/routes-ui/index.ts
@@ -2,7 +2,7 @@ export { Chain } from './Chain'
 export { ProvideChains } from './ChainContext'
 export { ChainSelect } from './ChainSelect'
 export { Connection, DirectConnection } from './Connection'
-export { routeId } from './routeId'
+export { getRouteId } from './getRouteId'
 export { Route, Routes } from './Routes'
 export { RouteSelect } from './RouteSelect'
 export { Waypoint, Waypoints } from './Waypoints'
