+++ deployables/app/app/routes.ts
@@ -49,7 +49,9 @@ export default [
           'routes/legacy-redirects/old-new-route-redirect.ts',
         ),
 
-        route('/account/:accountId', 'routes/account/edit.tsx'),
+        route('/account/:accountId', 'routes/account/edit.tsx', [
+          route(':routeId?', 'routes/account/routes.tsx'),
+        ]),
 
         ...prefix('/edit', [
           index('routes/edit/list-routes.tsx'),
