+++ deployables/app/app/routes/account/edit.spec.ts
@@ -408,4 +408,33 @@ describe('Edit account', () => {
       expect(activeRoute.route).toHaveProperty('waypoints', waypoints)
     })
   })
+
+  describe('Routes', () => {
+    it('lists all routes', async () => {
+      const tenant = await tenantFactory.create()
+      const user = await userFactory.create(tenant)
+
+      const wallet = await walletFactory.create(user)
+      const account = await accountFactory.create(tenant, user)
+
+      await routeFactory.create(account, wallet, {
+        label: 'Route A',
+      })
+      await routeFactory.create(account, wallet, {
+        label: 'Route B',
+      })
+
+      await render(href('/account/:accountId', { accountId: account.id }), {
+        user,
+        tenant,
+      })
+
+      expect(
+        await screen.findByRole('tab', { name: 'Route A' }),
+      ).toBeInTheDocument()
+      expect(
+        await screen.findByRole('tab', { name: 'Route B' }),
+      ).toBeInTheDocument()
+    })
+  })
 })
